+++ src/index.ts
@@ -1,19 +1,20 @@
 import postcss, { Root, Transformer } from 'postcss';
-import { PluginOptions } from '@types';
+import { PluginOptions, Autorename } from '@types';
 import { store, initStore } from '@data/store';
 import { parseKeyFrames, parseAtRules } from '@parsers/atrules';
 import { parseRules } from '@parsers/rules';
-import { appendRules, appendKeyFrames } from '@utilities/rules';
-export { PluginOptions, Mode, Source, PluginStringMap,  } from '@types';
+import { appendRules, appendKeyFrames, appendAutorenameRules } from '@utilities/rules';
+export { PluginOptions, Mode, Source, PluginStringMap, Autorename } from '@types';
 
 const transformer = (options: PluginOptions = {}): Transformer  => (
     (css: Root): void => {
-        initStore(options);        
+        initStore(options);     
         store.options.processKeyFrames && parseKeyFrames(css);
         parseAtRules(css);
         parseRules(css);
         appendRules();
-        appendKeyFrames();                 
+        appendKeyFrames();
+        store.options.autoRename !== Autorename.disabled && appendAutorenameRules();        
     }
 );