+++ src/utilities/rules.ts
@@ -1,4 +1,5 @@
 import { Rule, AtRule, Node } from 'postcss';
+import { ObjectWithProps, StringMap, Autorename } from '@types';
 import { COMMENT_TYPE, RTL_COMMENT_REGEXP, DECLARATION_TYPE, RULE_TYPE } from '@constants';
 import { store } from '@data/store';
 
@@ -46,4 +47,59 @@ export const appendKeyFrames = (): void => {
         atRule.after(atRuleFlipped);
         cleanRules(atRule, atRuleFlipped);
     }); 
+};
+
+export const appendAutorenameRules = (): void => {
+
+    const replaceHash: ObjectWithProps<string> = store.options.stringMap.reduce((hash: ObjectWithProps<string>, map: StringMap): ObjectWithProps<string> => {
+        const search = typeof map.search === 'string' ? [map.search] : map.search;
+        const replace = typeof map.replace === 'string' ? [map.replace] : map.replace;
+        search.forEach((s: string, index: number): void => {
+            hash[s] = replace[index];
+            hash[replace[index]] = s;
+        });
+        return hash;
+    }, {});
+    
+    const replaces = Object.keys(replaceHash).join('|');
+    const replaceRegExp = store.options.greedy
+        ? new RegExp(`(${replaces})`, 'g')
+        : new RegExp(`\\b(${replaces})\\b`, 'g');
+
+    const rulesHash: ObjectWithProps<boolean> = {};
+    const rulesToProcess: Rule[] = [];
+
+    store.rulesAutoRename.forEach((rule: Rule): void => {
+        let process = false;
+        rule.selectors.forEach((selector: string): void => {
+            if (replaceRegExp.test(selector)) {
+                rulesHash[selector] = true;
+                process = true;
+            }
+            replaceRegExp.lastIndex = 0;
+        });
+        if (process) {
+            rulesToProcess.push(rule);
+        }
+    });
+
+    rulesToProcess.forEach((rule: Rule): void => {
+
+        rule.selectors = rule.selectors.map((selector: string): string => {
+
+            const flip = selector.replace(replaceRegExp, (match: string, group: string): string => replaceHash[group]);
+            
+            if (
+                store.options.autoRename === Autorename.flexible ||
+                (store.options.autoRename === Autorename.strict && rulesHash[flip])
+            ) {
+                return flip;
+            }
+
+            return selector;
+
+        });
+
+    });
+
 };