+++ tests/override-basic-options.test.ts
@@ -4,7 +4,7 @@ import { readCSSFile } from './test-utils';
 
 const baseOptions: PluginOptions = {mode: Mode.override};
 
-describe('Override Tests', (): void => {
+describe('Override Tests Basic Options', (): void => {
 
   let input = '';
 
@@ -49,37 +49,6 @@ describe('Override Tests', (): void => {
 
   it('Override {useCalc: true}', (): void => {
     const options: PluginOptions = { ...baseOptions, useCalc: true };
-    const output = postcss([postcssRTLCSS(options)]).process(input);
-    expect(output.css).toMatchSnapshot();
-    expect(output.warnings()).toHaveLength(0);
-  });
-
-  it('Override custom ltrPrefix and rtlPrefix properties', (): void => {
-    const options: PluginOptions = { ...baseOptions, ltrPrefix: '.ltr', rtlPrefix: '.rtl' };
-    const output = postcss([postcssRTLCSS(options)]).process(input);
-    expect(output.css).toMatchSnapshot();
-    expect(output.warnings()).toHaveLength(0);
-  });
-
-  it('Override custom ltrPrefix and rtlPrefix properties as arrays', (): void => {
-    const options: PluginOptions = { ...baseOptions, ltrPrefix: ['.ltr', '.left-to-right'], rtlPrefix: ['.rtl', '.right-to-left'] };
-    const output = postcss([postcssRTLCSS(options)]).process(input);
-    expect(output.css).toMatchSnapshot();
-    expect(output.warnings()).toHaveLength(0);
-  });
-
-  it('Override custom ltrPrefix, rtlPrefix, and bothPrefix properties as arrays and processUrls: true', (): void => {
-    const options: PluginOptions = { ...baseOptions, ltrPrefix: ['.ltr', '.left-to-right'], rtlPrefix: ['.rtl', '.right-to-left'], bothPrefix: ['.ltr', '.left-to-right', '.rtl', '.right-to-left'], processUrls: true };
-    const output = postcss([postcssRTLCSS(options)]).process(input);
-    expect(output.css).toMatchSnapshot();
-    expect(output.warnings()).toHaveLength(0);
-  });
-
-  it('Override custom string map and processUrls: true', (): void => {
-    const stringMap: PluginOptions['stringMap'] = [
-      {search: 'left', replace: 'right'}
-    ];
-    const options: PluginOptions = { ...baseOptions, processUrls: true, stringMap };
     const output = postcss([postcssRTLCSS(options)]).process(input);
     expect(output.css).toMatchSnapshot();
     expect(output.warnings()).toHaveLength(0);